+++ src/app/order/components/process-order/process-order.component.ts
@@ -1,4 +1,9 @@
 import { Component, OnInit } from '@angular/core';
+import {Store} from "@ngrx/store";
+import {AppState} from "../../../store/app.reducer";
+import {selectOrderState} from "../../../store/order.selectors";
+import {map} from "rxjs/operators";
+import {Observable, of} from "rxjs";
 
 @Component({
   selector: 'app-process-order',
@@ -6,10 +11,14 @@ import { Component, OnInit } from '@angular/core';
   styleUrls: ['./process-order.component.scss']
 })
 export class ProcessOrderComponent implements OnInit {
+  totalSum: Observable<number> = of(0);
 
-  constructor() { }
+  constructor(private store: Store<AppState>) { }
 
   ngOnInit(): void {
+    this.totalSum = this.store.select(selectOrderState).pipe(
+      map(state => state.totalSum)
+    );
   }
 
 }