+++ src/app/product/components/product/product.component.ts
@@ -1,8 +1,7 @@
 import { Component } from '@angular/core';
-import {ProductModel} from "../../models/product.model";
-import {ProductRepository} from "../../services/product.repository";
-import {CartService} from "../../../cart/services/cart.service";
-import {SnackbarService} from "../../../core/services/snackbar.service";
+import {Store} from "@ngrx/store";
+import {AppState} from "../../../store/app.reducer";
+import {Observable, of} from "rxjs";
 
 @Component({
   selector: 'app-product',
@@ -10,14 +9,9 @@ import {SnackbarService} from "../../../core/services/snackbar.service";
   styleUrls: ['./product.component.scss']
 })
 export class ProductComponent {
-  constructor(
-    protected productRepository: ProductRepository,
-    protected cartService: CartService,
-    protected snackbarService: SnackbarService,
-  ) { }
+  showProgress: Observable<boolean> = of(false);
 
-  addProductToCart(item: ProductModel): void {
-    this.cartService.addItemToCart(item);
-    this.snackbarService.message('Product Added');
-  }
+  constructor(
+    protected store: Store<AppState>,
+  ) { }
 }