+++ src/app/product/product-routing.module.ts
@@ -2,8 +2,9 @@ import {RouterModule, Routes} from "@angular/router";
 import {NgModule} from "@angular/core";
 import {ProductListPageComponent} from "./components/product-list-page/product-list-page.component";
 import {ProductViewPageComponent} from "./components/product-view-page/product-view-page.component";
-import {resolveProductWithRedirect} from "../shared/guards/product.resolver";
+import {storeProductResolver} from "../shared/guards/product.resolver";
 import {ProductComponent} from "./components/product/product.component";
+import {fetchProduct} from "./store/product.actions";
 
 const routes: Routes = [
   {
@@ -13,7 +14,7 @@ const routes: Routes = [
       {
         path: ':productID',
         component: ProductViewPageComponent,
-        ...resolveProductWithRedirect('/products'),
+        ...storeProductResolver('products', fetchProduct),
       },
       { path: '', component: ProductListPageComponent },
     ],