+++ src/app/shared/guards/product.resolver.ts
@@ -1,18 +1,38 @@
-import {ActivatedRouteSnapshot, Resolve, Router, RouterStateSnapshot} from "@angular/router";
+import {ActivatedRouteSnapshot, Resolve, RouterStateSnapshot} from "@angular/router";
 import {ProductModel} from "../../product/models/product.model";
-import {ProductRepository} from "../../product/services/product.repository";
 import {Injectable} from "@angular/core";
 import {Observable} from "rxjs";
+import {Action, Store} from "@ngrx/store";
+import {AppState} from "../../store/app.reducer";
+import {take} from "rxjs/operators";
+import {selectSelectedProductAvailableByUrl, selectSelectedProductByUrl} from "../../store/product.selectors";
 
-export function resolveProductWithRedirect(redirectUrl: string) {
-  return {data: {redirectUrl}, resolve: {product: ProductResolver}};
+export function storeProductResolver(store: string, fetchAction: Action) {
+  return {
+    data: {
+      store,
+      fetchAction,
+    },
+    resolve: {product: ProductResolver}
+  };
 }
 
 @Injectable({providedIn: 'root'})
 export class ProductResolver implements Resolve<ProductModel> {
-  constructor(private productRepository: ProductRepository) {}
+  constructor(private store: Store<AppState>) {}
 
   resolve(route: ActivatedRouteSnapshot, state: RouterStateSnapshot): Observable<ProductModel> {
-    return this.productRepository.getProduct(+route.params['productID']);
+    let productAvailable = false;
+    this.store.select(selectSelectedProductAvailableByUrl)
+      .pipe(
+        take(1),
+      )
+      .subscribe((data => productAvailable = data));
+
+    if (!productAvailable) {
+      this.store.dispatch(route.data.fetchAction({productId: +route.params['productID']}));
+    }
+
+    return this.store.select(selectSelectedProductByUrl).pipe(take(1));
   }
 }